COMP: fix several conversion warnings.

+++ Code/ExternalCode/itkQt/itkQtProgressBar.cxx
@@ -92,7 +92,7 @@ QtProgressBar::ProcessEvent( itk::Object * caller,
              dynamic_cast< itk::ProcessObject *>( caller );
 
     const int tempvalue = static_cast<int>( 
-                        process->GetProgress() * this->maximum() );
+      process->GetProgress() * static_cast< float >( this->maximum() ) );
     std::cout << "New Value : " << tempvalue << std::endl;
     this->setValue( tempvalue );
     }
@@ -111,7 +111,7 @@ QtProgressBar::ConstProcessEvent( const itk::Object * caller,
               dynamic_cast< const itk::ProcessObject *>( caller );
 
     const int temp_value = static_cast<int>( 
-                        process->GetProgress() * this->maximum() );
+      process->GetProgress() * static_cast< float >( this->maximum() ) );
     std::cout << "New Value : " << temp_value << std::endl;
     this->setValue( temp_value );
     }
